REGRESSION(258946@main): utsname struct members as buffer fixed-length C arrays confuse StringView https://bugs.webkit.org/show_bug.cgi?id=250682
Reviewed by NOBODY (OOPS!).

In GLib's user-agent string construction, members of the utsname struct are
passed to makeString(). These are fixed-length C arrays, 64 in size, that
serve as buffers where uname() writes out the null-terminated string.

Because of that type, on a certain combination of compiler and standard library
the new StringView constructor handling fixed-length C arrays of characters can
end up being used, using the array length as the length of the observed string.
In this case this means that a length of 64 is being assumed, even when the
string itself is shorter than that.

The StringView constructor is still beneficial, but is now enhanced to assert
that the imposed fixed length matches the value of the strlen() call on the
character data.

For the two utsname struct members, StringView::fromLatin1() should be used
to avoid this.

* Source/WTF/wtf/text/StringView.h:
* Source/WebCore/platform/glib/UserAgentGLib.cpp:
(WebCore::platformVersionForUAString):